+++ frontend/src/lib/components/forms/edit-modal/InputPasswordField.svelte
