+++ frontend/src/routes/home/settings/suppliers/[id]/(components)/EditModal.svelte
@@ -2,8 +2,8 @@
 	import { enhance } from '$app/forms';
 	import { Dialog } from 'bits-ui';
 	import type { ActionData } from '../$types';
-	import InputTextField from '$lib/components/edit-modal-form/InputTextField.svelte';
-	import InputTextareaField from '$lib/components/edit-modal-form/InputTextareaField.svelte';
+	import InputTextField from '$lib/components/forms/edit-modal/InputTextField.svelte';
+	import InputTextareaField from '$lib/components/forms/edit-modal/InputTextareaField.svelte';
 
 	interface Props {
 		openState: boolean;
