+++ frontend/src/routes/home/settings/users/add/+page.svelte
@@ -1,9 +1,8 @@
 <script lang="ts">
 	import { enhance } from '$app/forms';
-	import InputPasswordField from '$lib/components/card-form/InputPasswordField.svelte';
-	import InputSelect from '$lib/components/card-form/InputSelect.svelte';
-	import InputTextField from '$lib/components/card-form/InputTextField.svelte';
-	import { redirect } from '@sveltejs/kit';
+	import InputPasswordField from '$lib/components/forms/card/InputPasswordField.svelte';
+	import InputSelect from '$lib/components/forms/card/InputSelect.svelte';
+	import InputTextField from '$lib/components/forms/card/InputTextField.svelte';
 	import toast from 'svelte-french-toast';
 
 	let { data, form } = $props();
